Moved IDrawable and Sprite to a new namespace
