+++ TeamProject/GlobalWarmingGame/GameObjectManager.cs
@@ -1,4 +1,5 @@
 ﻿using Engine;
+using Engine.Drawing;
 using GlobalWarmingGame.Interactions;
 using System.Collections.Generic;
 using System.Linq;
