+++ TeamProject/GlobalWarmingGame/Interactions/InteractableGameObject.cs
@@ -1,4 +1,5 @@
 ﻿using Engine;
+using Engine.Drawing;
 using GlobalWarmingGame.Action;
 using GlobalWarmingGame.ResourceItems;
 using Microsoft.Xna.Framework;
