+++ TeamProject/GlobalWarmingGame/Interactions/PassiveMovingInteractable.cs
@@ -1,4 +1,5 @@
 ﻿using Engine;
+using Engine.Drawing;
 using Engine.PathFinding;
 using GlobalWarmingGame.Action;
 using GlobalWarmingGame.ResourceItems;
